requirements.txt
Django==2.1.2
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-debug-toolbar==1.10.1
django-froala-editor==2.9.0
django-js-asset==1.1.0
django-select2==6.3.1
django-suit==0.2.26
django-taggit==0.23.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.7
sqlparse==0.2.4
